Array SD, Common SD, ConfigReader and config properties for stack
