+++ src/test/java/stepDefinition/Array_SD.java
@@ -17,14 +17,6 @@ public class Array_SD {
 	String Excelpath = ConfigReader.getexcelfilepath();
 	static String expectedMsg;
 
-	// @TS_array_02
-	@Given("The user is on the {string} after logged in")
-	public void the_user_is_on_the_after_logged_in(String pagename) {
-		String page_name = pagename.replaceAll("\\s+", "");
-		arrayPage.navigateTo(page_name);
-		Loggerload.info("The user is on the " + pagename + " after logged in");
-	}
-
 	@When("The user select Array item from the drop down menu")
 	public void the_user_select_array_item_from_the_drop_down_menu() {
 		arrayPage.dropdown_Array();
