+++ src/test/resources/config/config.properties
@@ -1,9 +1,23 @@
 browser=chrome
+excelfilepath=./src/test/resources/Exceldata/Login.xlsx
 url=https://dsportalapp.herokuapp.com
 homepage=https://dsportalapp.herokuapp.com/home
 loginpage=https://dsportalapp.herokuapp.com/login
-excelfilepath=./src/test/resources/Exceldata/Login.xlsx
-arrayurl=https://dsportalapp.herokuapp.com/array/
+
+#Array URLs
+Arraypage=https://dsportalapp.herokuapp.com/array/
+ArraysinPythonpage=https://dsportalapp.herokuapp.com/array/arrays-in-python/
+ArraysUsingList=https://dsportalapp.herokuapp.com/array/arrays-using-list/
+BasicOperationsinLists=https://dsportalapp.herokuapp.com/array/basic-operations-in-lists/
+ApplicationsofArray=https://dsportalapp.herokuapp.com/array/applications-of-array/
+Practicepage=https://dsportalapp.herokuapp.com/array/practice
+QuestionSearchthearray=https://dsportalapp.herokuapp.com/question/1
+QuestionMaxConsecutiveOnes=https://dsportalapp.herokuapp.com/question/2
+QuestionFindNumberswithEvenNumberofDigits=https://dsportalapp.herokuapp.com/question/3
+QuestionSquaresofaSortedArray=https://dsportalapp.herokuapp.com/question/4
+tryeditorurl=https://dsportalapp.herokuapp.com/tryEditor
+
+#Stack URLs
 Stackpage=https://dsportalapp.herokuapp.com/stack/
 OperationsinStackpage=https://dsportalapp.herokuapp.com/stack/operations-in-stack/
 StackImplementationpage=https://dsportalapp.herokuapp.com/stack/implementation/
